reader: Fix building of void and void pointer types
build_ir_node_for_void_pointer_type and build_ir_node_for_void_type
get the type IR from the environment, add it to the global scope of
the current translation unit and schedule it for type
canonicalization.

This patch change the code of these functions so that the adding to
the scope and the canonicalization are done only on the first
invocation of these functions.  Subsequent invocations just return the
type IR.

Otherwise, subsequent invocations were unnecessarily trying to add the
type IR to yet another global scope and were trying to canonicalize it
again.

	* src/abg-reader.cc (build_ir_node_for_void_pointer_type)
	(build_ir_node_for_void_type): Add the type to its scope and
	schedule it for canonicalization just once.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>