ctf-reader: Fix analyzing single kernel binaries
Analyzing a single kernel binary has been broken for a long time, it
seems.  For the record, I mean this:

    $ abidw --ctf vmlinux > vmlinux.ctf.abi

Doesn't work.

This patch fixes ctf::reader::slurp_elf_info to make it not crash on
binaries that have no CTF sections because there is a .ctfa file in
the same directory.

Also, it fixes ctf::reader::read_corpus to support loading the CTF
archive for individual kernel binaries and not just for entire trees
(corpus groups).  In the case of individual kernel binaries, the
archive can be built from the CTF sections of the binary.

	* src/abg-ctf-reader.cc (reader::slurp_elf_info): Do not crash on
	a kernel binary that has no CTF section because the CTF
	information is in a .ctfa file in the same directory.
	(reader::read_corpus): Support loading a CTF archive for a single
	kernel binary.  Also, support loading the CTF archive from the CTF
	sections in the binary itself not necessarily from a .ctfa file in
	the same directory.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>