tools-utils,btf-reader: Take modules into account for corpus group
It turns out the BTF front-end won't take kernel modules into account
when analyzing a linux kernel build tree.

This is because it doesn't support split and base BTF objects.

Please note that the BTF API is not yet documented in
https://docs.kernel.org/bpf/btf.html, so my understanding of how it
works comes the source code of the bpftool program available at
https://github.com/libbpf/bpftool and https://github.com/libbpf/libbpf.

When analyzing the vmlinux binary, the btf__parse function returns a
"base BTF object" which contains the type information contained in the
.BTF section of that binary.  The memory of the BTF types describing
the types of vmlinux lives in the base BTF object.  So the base BTF
object must be kept around for the entire time of the analysis of the
Linux kernel tree (vmlinux + modules).

When analyzing a kernel module however, it's the btf__parse_split
function that is used.  That function takes into parameter the "base
BTF object" representing the type information of the corresponding
vmlinux binary and returns a "split BTF object" which contains the
type information contained in the .BTF section of the kernel module.
Please note that the split BTF object does reference the base BTF
object as well.  The memory of the BTF types describing the types the
kernel module lives in the split BTF object.  Because some types and
decls might be exported by a kernel module and used by another one,
the memory of a given split BTF object might be needed to be kept
alive for the entire time of the analysis of the Linux kernel tree as
well.

It's this base/split model of BTF that is not supported by the BTF
front-end.

This patch introduces support for that model.

At any point in time, the BTF front-end now has one handle to the base
BTF object and one handle to the current split BTF object if we are
looking at kernel module.  The base and split BTF objects are kept
around for the entire lifetime of the front-end.

For a given binary, each type is identified by a type ID.  The
validity of a given type ID is effective only for that binary.  The
same type ID in another binary might designate another type.  So
rather than maintaining a map that associates a type ID with a libabigail
IR artifact, the front-end now maintains a map that associates a
*type* with libabigail IR artifact.

With this changes, abidw can now analyze and emit the ABIXML for a
linux kernel tree in 28 seconds, using BTF.  It can self-compare a
kernel tree in 36 seconds.

	* src/abg-btf-reader.cc (btf_type_to_abi_artifact_map_type):
	Rename btf_type_id_to_abi_artifact_map_type into this.  Make it
	associate a btf_type* to a libabigail IR, rather than a type ID to
	a libabigail IR.
	(reader::{base_btf_handle_, base_btf_file_name_,
	split_btf_handle_, split_btf_file_name_, split_btfs_to_free_}):
	Define new data members.
	(reader::{btf_handle_}): Remove.
	(reader::base_btf_handle): Renamed btf_handle into this.  Create
	the BTF object using btf__parse.  Store the name of the binary
	into reader::base_btf_file_name_.
	(reader::read_split_btf): Define new member function.
	(reader::btf_handle): This is new code for an existing member
	function name.  Return the current split BTF object handle if
	applicable, otherwise create the base BTF object handle.
	(reader::btf_type_to_artifacts_): Renamed
	btf_type_id_to_artifacts into this.
	(reader::associate_artifact_to_btf_type): Renamed
	associate_artifact_to_btf_type_id into this.
	(reader::schedule_type_for_canonicalization): Do not schedule a
	type that has already been canonicalized.
	(reader::nr_btf_types): Take a BTF* parameter.
	(reader::~reader): Free all the allocated split BTF objects.
	(reader::read_debug_info_into_corpus): If we are looking at a
	split BTF, then set the first type ID as the one right after the
	last type ID of the base BTF.
	(reader::build_ir_node_from_btf_type): Clean-up the logic for
	building type void type.
	(reader::{build_ir_node_for_void_type,
	build_ir_node_for_void_pointer_type,
	build_ir_node_for_variadic_parameter_type,
	build_enum_underlying_type, build_array_type}): Schedule type for
	canonicalization at the end of the IR construction, not right
	away.
	* src/abg-tools-utils.cc (create_best_elf_based_reader): Pass
	show_all_types and linux_kernel_mode arguments to
	btf::create_reader.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>