ctf-reader: Enumerate dicts in the archive rather than using their name
ctf::reader::process_ctf_archive calls ctf_dict_open with the name of
the (parent) dictionary to open.  If that parent dictionary is set to
an unexpected name, then the call to ctf_dict_open fails.  This can
happen for instance when at link time, the name of the parent
dictionary is set to an arbitrary name using the
"ctf_link_set_memb_name_changer" function.

This patch enumerates the dictionaries of the archive to avoid having
to know the name of the parent dictionary.  The enumeration is done
using the ctf_archive_next function of libctf.

There is currently no binary with an unexpected dictionary name in the
test suite so this patch cannot be tested for that particular case.
I'd be glad to have such binaries added to the test suite.

In the mean time this patch has been tested successfully using "make
fullcheck" on the existing test suite.

	* src/abg-ctf-reader.cc (reader::process_ctf_archive):  Do not use
	ctf_dict_open to open the dictionary by name.  Rather, enumerate
	the dictionaries of the current archive by using ctf_archive_next
	just like what lookup_symbol_in_ctf_archive does.  Set the
	argument of the skip_parent parameter to "false" to ensure we get
	the parent dictionary.
	(lookup_symbol_in_ctf_archive): Clean this up to properly
	initialize the parameters and to document the arguments to the
	parameters of ctf_archive_next.  Use nullptr instead of NULL.
	* tests/data/test-diff-pkg-ctf/test-rpm-report-1.txt: Adjust.
	* tests/data/test-diff-pkg-ctf/test-rpm-report-2.txt: Adjust.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>