reader: Avoid duplicating recursive types
Sometimes, building a sub-type of a complex type can trigger the
building of the complex type itself.  This is in the case of a
recursive complex type where one of its sub-types has the complex type
as a sub-type.

In those cases, the complex type should not be duplicated.  This patch
ensures that.

	* src/abg-reader.cc (build_qualified_type_decl)
	(build_pointer_type_def, build_reference_type_def)
	(build_ptr_to_mbr_type, build_subrange_type, build_array_type_def)
	(build_enum_type_decl, build_typedef_decl): After a sub-type is
	built, check if the complex type we are looking at is built too.
	If it is, then return it.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>