ir: Fix name setting of a ptr-to-mbr-type
The name of a pointer to member type is currently set at the time of
construction of the type.

This can be problematic because at the time of construction of the
type, the member type might not be yet fully constructed.  That would
result in a name (for the pointer to member type) that is incorrect
and stays incorrect for the lifetime of the ptr-to-mbr-type.

To be correct, the name of the pointer to member type must always
reflect the current state of the member and containing types.

This patch does just that.

The patch turns decl_base::get_name into a virtual member function.
It then creates a ptr_to_mbr_type::get_name virtual member function
that dynamically overloads decl_base::get_name.  Then,
ptr_to_mbr_type::get_name just builds and returns the qualified name
of the pointer to member type.

	* include/abg-ir.h (decl_base::{g,s}et_name): Make these member
	functions virtual.
	(ptr_to_mbr_type::get_name): Add new virtual overload.
	* src/abg-ir.cc (ptr_to_mbr_type::ptr_to_mbr_type): Do not set the
	name of the type here.  Just say that it's not anonymous.
	(ptr_to_mbr_type::get_name): Define new (virtual) member function.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>