dwarf-reader: Support LLVM's lingo of declaration-ness
In the DWARF emitted by LLVM, a type DIE can have the
DW_AT_declaration attribute present (saying that it's a declaration)
and yet have the DW_AT_byte_size present, implying that the type is
defined, at the same time.  OK, in all fairness, a definition is a
declaration too, but then, the DW_AT_declaration attribute becomes a
tad useless on a definition.

Note that in the DWARF emitted by GCC, when a DIE has the
DW_AT_declaration set, if no size attribute is present then Libabigail
considers that it's a declaration, not a definition.

This state of things makes libabigail wrongly think that some types
(especially some enums) are defined, when they are just declarations.
And that leads to self-comparison issues down the road.

For instance, consider this DIE sub-tree from the libLLVM.so of
llvm-libs-17.0.6-5.el9.x86_64:

 [1745df]    namespace            abbrev: 46
             name                 (GNU_strp_alt) "llvm"
             sibling              (ref_udata) [1745fe]
 [1745e5]      class_type           abbrev: 98
               name                 (GNU_strp_alt) "raw_ostream"
               declaration          (flag_present) yes
[...]

 [1745f6]        enumeration_type     abbrev: 116
                 name                 (GNU_strp_alt) "Colors"
                 byte_size            (data1) 4
                 declaration          (flag_present) yes

This "enum Colors" DIE here is a just a declaration, even though it
has a DW_AT_byte_size attribute set.  We see that it has no children
DIE.

But then, later, we can see this DIE sub-tree:

 [484843]    namespace            abbrev: 39
             name                 (GNU_strp_alt) "llvm"
 [484848]      class_type           abbrev: 90
               containing_type      (ref_udata) [484848]
               calling_convention   (data1) pass_by_reference (4)
               name                 (GNU_strp_alt) "raw_ostream"
               byte_size            (data1) 64
               decl_file            (data1) raw_ostream.h (1)
               decl_line            (data1) 52
[...]
 [4848b9]        enumeration_type     abbrev: 125
                 name                 (GNU_strp_alt) "Colors"
                 byte_size            (data1) 4
                 declaration          (flag_present) yes
 [4848bf]        member               abbrev: 108
                 name                 (GNU_strp_alt) "RED"
                 type                 (ref_udata) [485234]
                 decl_file            (data1) raw_ostream.h (1)
                 decl_line            (data1) 110
                 external             (flag_present) yes
                 declaration          (flag_present) yes
                 accessibility        (data1) public (1)
                 const_value          (sdata) 1 (1)
 [4848ca]        member               abbrev: 108
                 name                 (GNU_strp_alt) "GREEN"
                 type                 (ref_udata) [485234]
                 decl_file            (data1) raw_ostream.h (1)
                 decl_line            (data1) 111
                 external             (flag_present) yes
                 declaration          (flag_present) yes
                 accessibility        (data1) public (1)
                 const_value          (sdata) 2 (2)
[...]

Here, we see the actual definition of the "enum Colors" represented
before by DIE 0x1745f6.  The definition is not linked in any way to
the previous declaration.  So the DWARF reader now thinks that both
enum Colors (at DIE 0x1745f6 and 0x4848b9) are two different enums
with the same name.  And of course, that can lead to self-comparison
issues down the road.

In this patch, the DWARF reader is instructed to detect that DIE
0x1745f6 is a declaration, by noticing that it has no children DIE.

Once that was done, I noticed that the pass that resolves declarations
to their definitions is broken.

The pass is broken because declarations are schedule to be processed
by that pass too early, before they got added to their lexical scope.
At that point, the declarations have non-qualified names.  When the
declaration-resolution time comes (after the ABI corpus is fully
constructed), the definitions of these declarations are added to their
lexical scope and thus have fully-qualified names.  So the
declaration-resolution pass looks for definitions by looking up their
non-qualified names (as the declarations had non-qualified names, due
to their early scheduling), and it never finds any because they all
have full-qualified names!

So the patch fixes this by deferring the scheduling of declarations to
the declaration-resolution pass, for after they are added to their
lexical scope.

	* src/abg-dwarf-reader.cc (die_is_declaration_only): A DIE is a
	declaration if it has the DW_AT_declaration property and if it
	either has no size attribute or no children DIE.
	(reader:: maybe_schedule_decl_only_type_for_resolution): Define
	new function.
	(build_enum_type, add_or_update_class_type)
	(add_or_update_union_type): Do not schedule a declaration for
	declaration-to-definition-resolution pass here.
	(build_typedef_type): If the typedef is a naming typedef, schedule
	the underlying type for the declaration-to-definition-resolution
	pass here.
	(build_ir_node_from_die): Schedule the declaration for
	declaration-to-definition-resolution pass here, after it's been
	added to its scope.
	* tests/data/test-diff-filter/test41-report-0.txt: Adjust.
	* tests/data/test-read-dwarf/PR22015-libboost_iostreams.so.abi:
	Likewise.
	* tests/data/test-read-dwarf/PR22122-libftdc.so.abi: Likewise.
	* tests/data/test-read-dwarf/test-libandroid.so.abi: Likewise.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>