{btf,ctf,dwarf}-reader, ir: Fix self-comparison debugging for corpus groups
When debugging the self comparison of a corpus group, libabigail fails
to detect that two ABI artifacts belong to the same corpus group.
This is because it only takes into account each individual corpus, not
the fact that several corpora can be grouped into one corpus group.

This patch introduces a new function type_originates_from_corpus which
can detect that a type originates from a corpus or a from a corpus
group.  That function is used to detect that the type being
canonicalized belongs to the ABIXML corpus group and so its canonical
type must match the one of its originating ELF type.

	* src/abg-btf-reader.cc (reader::read_debug_info_into_corpus):
	When reading an ELF binary, which is part of a corpus group, for
	the purpose of self-comparison, set that corpus group as the
	comparison input.  Otherwise, if the ELF binary is not part of the
	corpus group, set its corpus as the comparison input.  This is
	done only when we are debugging the self-comparison process by
	doing abidw --debug-abidiff.
	* src/abg-ctf-reader.cc (reader::read_corpus): Likewise.
	* src/abg-dwarf-reader.cc (reader::read_debug_info_into_corpus):
	Likewise.
	* src/abg-ir-priv.h (type_originates_from_corpus): Declare ...
	* src/abg-ir.cc (type_originates_from_corpus): ... new function.
	(type_base::get_canonical_type_for): When debugging the
	self-comparison process, use the new type_originates_from_corpus
	to know if the current type being canonicalized originates from
	the second corpus.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>