dwarf-reader: Fix support of suppression specifications
While looking at something else, I realized that when functions and
variables are to be dropped on the floor by a suppression
specification, there are time where the function is suppressed but not
dropped from the ABI corpus.  This is due to some thinkos in the DWARF
reader code.  Fixed thus.

	* src/abg-dwarf-reader.cc (function_is_suppressed)
	(variable_is_suppressed): Do not return too early when the
	function doesn't seem suppressed, yet.
	* tests/data/test-diff-suppr/test31-report-0.txt: Adjust.
	* tests/data/test-diff-suppr/test32-report-1.txt: Adjust.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>