dwarf-reader,tools-utils: Add statistics about built/suppressed functions
When running abidw --verbose, it's useful to have information about
the number of functions and variable that were actually constructed or
suppressed.  This is useful to see how effective a suppression
specification that instructs libabigail to drop certains
functions/variables actually is.

This patch emits statistics about the amount of functions/variables
represented in the IR and dropped on the floor.

	* src/abg-dwarf-reader.cc (reader::stats): Define new struct.
	(reader::stats_): Define a new data member of type stats.
	(build_or_get_fn_decl_if_not_suppressed)
	(build_or_get_var_decl_if_not_suppressed): Increments the stats.
	(reader::read_debug_info_into_corpus): Emit the stats about number
	of suppressed/kept functions and variables at the end of the
	corpus building.
	* src/abg-tools-utils.cc (load_vmlinux_corpus): Emit the status
	about the number of functions/variables in the corpus.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>