dwarf-reader,ir: Fix endless loop while analyzing DWARF from Modula-2
While reading the DWARF of libm2iso.so.19.0.0 from the package
libgm2-14.2.1-1.fc40.x86_64.rpm (Modula-2 library for the gcc-gm2
package), abidw crashes due to a stack overflow, because of an endless
loop while trying to print the type names for DIEs in the DWARF
reader.

This is due to this kind of recursive construct of an anonymous type
at offset 0xf1:

 [    f1]    structure_type       abbrev: 30
             byte_size            (data1) 32
             decl_file            (data1) RTentity.mod (12)
             decl_line            (data1) 34
             decl_column          (data1) 23
             sibling              (ref_udata) [   126]
 [    f7]      member               abbrev: 64
               name                 (GNU_strp_alt) "left"
               decl_file            (data1) RTentity.mod (12)
               decl_line            (data1) 35
               decl_column          (data1) 26
               type                 (ref_udata) [   126]
               data_member_location (data1) 0
[...]

Consider the data member name "left" at offset 0xf7.  The offset of
it's type DIE is 0x126.

Here is the DIE at offset 0x126:

 [   126]    pointer_type         abbrev: 48
             byte_size            (implicit_const) 8
             type                 (ref_udata) [    f1]

It's a pointer to a type described by the DIE at offset ... 0xf1.
That is a cycle in the graph of that anonymous type, right there.

To name an anonymous struct like this, the DWARF reader constructs its
"flat representation".  Namely, the idea is to represent the struct by
a string that looks like:

    "struct {type1 data1, type2 data2}"

For that, the DWARF reader walks the DIE tree and emit the flat
representation recursively.

But here, because the DIE tree contains a cycle, the code goes into an
endless loop.

There is a similar issue with function types

Note that later, when the libabigail IR is built, we face a similar
problem.

This patch introduces cycle detection & avoidance in the code that
emit type names in the DWARF reader and in the Libabigail IR.  It adds
the libgm2 RPM to the testsuite and self-compares it using abipkgdiff,
in the test-diff-pkg test harness.

	* src/abg-dwarf-reader.cc (die_qualified_type_name)
	(die_qualified_decl_name, die_qualified_type_name_empty)
	(die_function_signature, die_pretty_print_type)
	(die_pretty_print_decl, die_pretty_print)
	(reader::{get_die_qualified_type_name}): Take a new guard
	parameter.  Pass it to all the type name construction function,
	all the way down to die_class_flat_representation and
	die_return_and_parm_names_from_fn_type_die.
	(die_qualified_name, die_type_name)
	(die_class_or_enum_flat_representation)
	(die_class_or_enum_flat_representation)
	(reader::{get_die_pretty_type_representation,
	get_die_pretty_representation}): Add an overload that takes a
	guard parameter and rewrite the previous overload in terms of the
	new one.
	(die_class_flat_representation)
	(die_return_and_parm_names_from_fn_type_die): Take a new guard
	parameter.  Use it to detect and avoid cycles during the
	construction of the type name.
	* src/abg-ir-priv.h
	(class_or_union::priv::is_printing_flat_representation_): Define
	new data member.
	(class_or_union::priv::{set,unset,is}_printing_flat_representation):
	Define member functions.
	(function_type::priv::is_pretty_printing_): Define new data
	member.
	(function_type::priv::{set,unset,}_is_pretty_printing): Define
	member functions.
	* src/abg-ir.cc (get_class_or_union_flat_representation)
	(add_outer_pointer_to_fn_type_expr): Implement cycle detection and
	avoidance for recursive classes and function types.
	* tests/data/test-diff-pkg/libgm2-14.2.1-1.fc40.x86_64-self-check-report.txt:
	New test reference output.
	* tests/data/test-diff-pkg/libgm2-14.2.1-1.fc40.x86_64.rpm: New
	input RPM.
	* tests/data/test-read-dwarf/test-libandroid.so.abi: Adjust.
	* tests/data/Makefile.am: Add the new test material to source
	distribution.
	* tests/test-diff-pkg.cc (in_out_specs): Add the new test inputs
	to the test harness.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>