dwarf-reader,reader.cc: Fix function virtuality setting
When setting the virtual-ness of a function (whether it's virtual or
not and its vtable offset), one must set the vtable offset before
setting the is_virtual property because setting the later triggers a
fix-up that uses the former.

It turns out some parts of the code got the order wrong.  To fix that,
this patch introduce a new set_member_function_virtuality that ensures
things are done in the proper order.

SET_member_function_vtable_offset and set_member_function_is_virtual
are now static functions that are used only by the new
set_member_function_virtuality.

	* include/abg-fwd.h (set_member_function_vtable_offset)
	(set_member_function_is_virtual): Remove these function
	declarations.
	(set_member_function_virtuality): Declare new function.
	* include/abg-ir.h (decl_base::get_context_rel): Make this public.
	(set_member_function_is_virtual)
	(set_member_function_vtable_offset): Remove these friend
	declaration for the decl_base class.
	* src/abg-dwarf-reader.cc (finish_member_function_reading): Use
	the new set_member_function_virtuality in lieu of the previous
	set_member_function_is_virtual and
	set_member_function_vtable_offset.
	* src/abg-ir.cc (set_member_function_vtable_offset)
	(set_member_function_is_virtual): Make these functions static.
	(set_member_function_virtuality): Define new functions.
	(class_or_union::add_member_function): Use the new
	set_member_function_virtuality in lieu of
	set_member_function_is_virtual and
	set_member_function_vtable_offset.
	* src/abg-reader.cc (build_class_decl): Likewise.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>