dwarf-reader: Do not fix ELF symbols for virtual destructors
There are cases in DWARF emitted by old producers where some member
function DIEs lack reference to the ELF symbol they relate to.  To
address that issue the DWARF reader sets some functions (with linkage
name and no ELF symbol) to the ELF symbol matching that linkage name.

We should not do that destructors however, because of false positives
and because it seems destructors are well handled even by the old
producers that I had access to.

This finishes to fix the self-comparison of the llvm-libs package of fc37.

	* src/abg-dwarf-reader.cc (finish_member_function_reading): Do not
	schedule virtual destructors for ELF symbol-related function
	fixup.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>