comparison: Sort anonymous types using their flat representation
When changes to types non-reachable from external interfaces are
emitted, they are sorted using the pretty representation of said
types.  In the current form, it's the "internal" pretty representation
that is used for the sorting.  That internal form, which is normally
used for the purpose of type canonicalization, doesn't differentiate
an anonymous types against another one.  For instance, all anonymous
enums are named __anonymous_enum__.  In that case, the sorting of
anonymous type changes becomes non stable.

This patch uses the non-internal form of type pretty representation
for the sorting.  In that form, anonymous enums are named using their
flat representation.

	* src/abg-comparison-priv.h (diff_comp::operator()): Use the
	non-internal pretty representation for sorting.
	* tests/data/test-abidiff-exit/test-anon-types-report-1.txt:
	Adjust.
	* tests/data/test-abidiff-exit/test-anonymous-enums-change-report-v1.txt:
	Adjust.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>