ir: Use definition of decl-only parm type in function type names
When the type of a function parameter is a declaration-only type and
if it has a definition, then this patch uses that definition in the
representation of the function type name.

Note that the ABIXML writer always emits definitions of
declaration-only types when they are available so if we don't use
those definitions in the function type representations, there might be
subtle self-comparison errors down the road.

This patch fixes the self-comparison of the OpenMolcas package in
Fedora 36 which was failing for the command:

    $ fedabipkgdiff --self-compare -a --from fc36 OpenMolcas

	* src/abg-ir.cc (stream_pretty_representation_of_fn_parms): Use
	the definition of decl-only function parameter types in the
	representation.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>