ir: Always use canonical types in comparison when possible
After staring at DWARF and ABIXML dumps for a while, I realized that
in the current (non-perfect) state of things, comparing type A and
type B can be slightly different from comparing canonical_type_of(A)
against type B, for instance.  This is essentially because comparing
canonical_type_of(A) against B rather than A against B changes the
order in which the nodes of the graph of types are visited.  Because
that graph has cycles, the order of visiting would essentially change
the value of the hashes that we compute during those visits.

This in turn changes the value of the comparisons depending on the
order of the comparisons.

So, to avoid those subtle changes, this patch ensures that whenever a
type has a canonical type, then it's that canonical type that is
always used in comparison.  This ensures that types coming from ABIXML
(as they are all canonical types) are compared in the same order as
types coming from the ELF binaries.

This fixes the self-comparison of the infinipath-psm package from
fc36, referenced in bug https://sourceware.org/bugzilla/show_bug.cgi?id=29413.

	* src/abg-ir.cc (maybe_get_canonical_type): Define new function
	template.
	(try_canonical_compare): Use the above function to always use
	canonical types in comparison.
	* tests/data/test-diff-pkg/infinipath-psm-3.3-26_g604758e_open.6.fc36.5.x86_64-self-check-report.txt:
	New test reference output.
	* tests/data/test-diff-pkg/infinipath-psm-3.3-26_g604758e_open.6.fc36.5.x86_64.rpm:
	Add new binary input.
	* tests/data/test-diff-pkg/infinipath-psm-debuginfo-3.3-26_g604758e_open.6.fc36.5.x86_64.rpm:
	Likewise.
	* tests/data/Makefile.am: Add the new test material above to
	source distribution.

Signed-off-by: Dodji Seketeli <dodji@redhat.com>